Move media type filter types and helpers to owned files
Types live in types/; path and filter helpers in mediaHelpers. Tests
reuse a shared allowedFileExtensions fixture from nuxt-aligned lists.
