+++ utils/mediaHelpers.ts
@@ -1,3 +1,9 @@
+import type {
+  AllowedFileExtensions,
+  GalleryMediaType,
+  MediaTypeFilterValue,
+} from "@/types";
+
 /**
  * Extracts the share ID from a Filebrowser share URL or returns the input if it's already a hash.
  * @example
@@ -148,3 +154,109 @@ export const getBaseUrlFromInput = (
 
   return defaultBaseUrl;
 };
+
+const normalizeExtension = (ext: string): string =>
+  ext.toLowerCase().replace(/^\./, "");
+
+const fileNameHasExtension = (fileName: string, ext: string): boolean => {
+  const suffix = normalizeExtension(ext);
+  return fileName.toLowerCase().endsWith(`.${suffix}`);
+};
+
+const extensionListIncludes = (list: string[] = [], ext: string): boolean =>
+  list.some((candidate) => normalizeExtension(candidate) === ext);
+
+/** Extracts file paths with valid extensions from a feature object. */
+export const getFilePathsWithExtension = (
+  feature: { [key: string]: unknown },
+  allExtensions: { [category: string]: string[] },
+  mediaColumn?: string,
+): string[] => {
+  if (!feature) return [];
+
+  const filePaths: string[] = [];
+  const keysToProcess = mediaColumn ? [mediaColumn] : Object.keys(feature);
+
+  keysToProcess.forEach((key) => {
+    if (typeof feature[key] !== "string") return;
+    if (feature[key].includes("attachment")) return;
+
+    const files = feature[key].split(",");
+    // handle ["\"5bf52de27e1a7b36f2d2cec254b766c8.jpg\""]
+
+    files.forEach((file: string) => {
+      const cleanedFile = file
+        .trim()
+        .replace(/^[\s"'\\[]+|[\s"'\\[\]]+$/g, "") // Remove brackets, quotes, backslashes, and whitespace from edges
+        .replace(/ /g, "_");
+
+      const hasValidExtension = Object.values(allExtensions).some(
+        (extensions) =>
+          extensions.some((ext: string) =>
+            fileNameHasExtension(cleanedFile, ext),
+          ),
+      );
+
+      if (hasValidExtension) {
+        filePaths.push(cleanedFile);
+      }
+    });
+  });
+
+  return filePaths;
+};
+
+export const getMediaTypesForEntry = (
+  feature: { [key: string]: unknown },
+  allExtensions: AllowedFileExtensions,
+  mediaColumn?: string,
+): GalleryMediaType[] => {
+  const paths = getFilePathsWithExtension(feature, allExtensions, mediaColumn);
+  const types = new Set<GalleryMediaType>();
+
+  for (const path of paths) {
+    const ext = path.split(".").pop()?.toLowerCase() ?? "";
+    if (!ext) continue;
+    if (extensionListIncludes(allExtensions.audio, ext)) types.add("audio");
+    if (extensionListIncludes(allExtensions.image, ext)) types.add("image");
+    if (extensionListIncludes(allExtensions.video, ext)) types.add("video");
+  }
+
+  return Array.from(types);
+};
+
+export const getMediaTypeFilterOptions = <T extends { [key: string]: unknown }>(
+  items: T[],
+  allExtensions: AllowedFileExtensions,
+  mediaColumn?: string,
+): MediaTypeFilterValue[] => {
+  const options: MediaTypeFilterValue[] = [];
+  if (allExtensions.image?.length) options.push("image");
+  if (allExtensions.audio?.length) options.push("audio");
+  if (allExtensions.video?.length) options.push("video");
+
+  const hasNone = items.some(
+    (item) =>
+      getMediaTypesForEntry(item, allExtensions, mediaColumn).length === 0,
+  );
+  if (hasNone) options.push("none");
+
+  return options;
+};
+
+export const filterByMediaTypes = <T extends { [key: string]: unknown }>(
+  items: T[],
+  selectedTypes: MediaTypeFilterValue[],
+  allExtensions: AllowedFileExtensions,
+  mediaColumn?: string,
+): T[] => {
+  if (!selectedTypes.length) return items;
+  return items.filter((item) => {
+    const present = getMediaTypesForEntry(item, allExtensions, mediaColumn);
+    return selectedTypes.some((type) => {
+      if (type === "none") return present.length === 0;
+      return present.includes(type);
+    });
+  });
+};
+
